+++ src/static/css/header.css
@@ -1,5 +1,5 @@
 .header{
-    background-image: url(../pictures/ai_header.jpg);
+    background-image: url(../img/ai_header.jpg);
     background-repeat: no-repeat;
     background-color: 	#001123;
     height: 378px;
